TableHeader, TableBody 컴포넌트 분리
